Introduced tag-checking.
There's now a "tagstack" that will keep track of any opened/closed html tags,
allowing detection when they don't add up.
(although the error messages for that are a bit cryptic for now …)
This only works with tags that were opened/closed with the \otag or \ctag command,
though, leading to some confusion with the (for now) hardcoded paragraphs/html body
elements.